+++ app/app.py
@@ -8,9 +8,8 @@ from app.routers import (
     static_routers, user_routers, collection_routers, document_routers,
     system_routers)
 from app.database import Base, sessionmanager, get_session
-from app.errors import Msg
+from app.errors import SERVER_ERROR
 from contextlib import asynccontextmanager
-from pydantic import ValidationError
 from time import time
 from uuid import uuid4
 import os
@@ -107,20 +106,8 @@ async def exception_handler(request: Request, e: Exception):
     # ctx = get_context()
     elapsed_time = time() - ctx.request_start_time
 
-    if isinstance(e, ValidationError):
-        status_code = status.HTTP_422_UNPROCESSABLE_ENTITY
-        error_details = e.errors()
-
-    elif isinstance(e, (FileNotFoundError, FileExistsError)):
-        status_code = status.HTTP_400_BAD_REQUEST,
-        error_details = {"detail": Msg.BAD_REQUEST}
-
-    else:
-        status_code = status.HTTP_500_INTERNAL_SERVER_ERROR
-        error_details = {"detail": Msg.SERVER_ERROR}
-
     log.error("Request failed; module=app; function=exception_handler; "
-              "elapsed_time=%s; status_code=%s; e=%s;" % (
-                  elapsed_time, status_code, str(e)))
-    response_content = jsonable_encoder({"detail": error_details})
-    return JSONResponse(status_code=status_code, content=response_content)
+              "elapsed_time=%s; e=%s;" % (elapsed_time, str(e)), exc_info=True)
+
+    return JSONResponse(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+                        content=jsonable_encoder({"detail": SERVER_ERROR}))