+++ app/routers/document_routers.py
@@ -51,10 +51,12 @@ async def document_upload(
     collection_repository = Repository(session, cache, Collection)
     collection = await collection_repository.select(id=schema.collection_id)
     if not collection:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
+        raise E("collection_id", schema.collection_id, E.RESOURCE_NOT_FOUND,  # noqa E501
+                status_code=status.HTTP_404_NOT_FOUND)
 
     elif collection.is_locked:
-        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+        raise E("collection_id", schema.collection_id, E.RESOURCE_LOCKED,
+                status_code=status.HTTP_423_LOCKED)
 
     # Save uploaded file
     filename = str(uuid.uuid4()) + cfg.DOCUMENTS_EXTENSION
@@ -168,7 +170,8 @@ async def document_select(
     document = await document_repository.select(id=schema.document_id)
 
     if not document:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
+        raise E("document_id", schema.document_id, E.RESOURCE_NOT_FOUND,
+                status_code=status.HTTP_404_NOT_FOUND)
 
     hook = Hook(session, cache, request, current_user=current_user)
     await hook.execute(H.AFTER_DOCUMENT_SELECT, document)
@@ -188,11 +191,11 @@ async def document_update(
     document_repository = Repository(session, cache, Document)
     document = await document_repository.select(id=schema.document_id)
     if not document:
-        raise E("document_id", schema.document_id, E.ENTITY_NOT_FOUND,
+        raise E("document_id", schema.document_id, E.RESOURCE_NOT_FOUND,
                 status_code=status.HTTP_404_NOT_FOUND)
 
     elif document.document_collection.is_locked:
-        raise E("document_id", schema.document_id, E.ENTITY_LOCKED,
+        raise E("document_id", schema.document_id, E.RESOURCE_LOCKED,
                 status_code=status.HTTP_423_LOCKED)
 
     if document.document_collection.id != schema.collection_id:
@@ -201,11 +204,11 @@ async def document_update(
             id=schema.collection_id)
 
         if not collection:
-            raise E("collection_id", schema.collection_id, E.ENTITY_NOT_FOUND,
+            raise E("collection_id", schema.collection_id, E.RESOURCE_NOT_FOUND,  # noqa E501
                     status_code=status.HTTP_404_NOT_FOUND)
 
         if collection.is_locked:
-            raise E("collection_id", schema.collection_id, E.ENTITY_LOCKED,
+            raise E("collection_id", schema.collection_id, E.RESOURCE_LOCKED,  # noqa E501
                     status_code=status.HTTP_423_LOCKED)
 
     document.collection_id = schema.collection_id
@@ -240,10 +243,12 @@ async def document_delete(
 
     document = await document_repository.select(id=schema.document_id)
     if not document:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
+        raise E("document_id", schema.document_id, E.RESOURCE_NOT_FOUND,
+                status_code=status.HTTP_404_NOT_FOUND)
 
     elif document.document_collection.is_locked:
-        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+        raise E("document_id", schema.document_id, E.RESOURCE_FORBIDDEN,
+                status_code=status.HTTP_403_FORBIDDEN)
 
     if document.comments_count > 0:
         # TODO: delete related comments